+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 #### Features
   - The load log file can now be configured to include relations (#94)
   - The load log file can now be configured to include skipped objects (#148)
+  - Detection of duplicate UUIDs (#185)
 
 #### Bugfixes
   - Fixed an access violation which could occur when loading from CSV or SQL auxilliary tables (#92)